Optimize/fix bugs in ConcurrentLinkedList and region manager changes
ChunkRegionManager needs to be changed to always merge neighbour regions,
as modifying vanilla code to handle ticking regions is not maintainable.

ConcurrentLinkedList now beats CLQ in performance.

fastClear in ConcurrentLinkedList has been removed as a race condition
(in poll()) can set the head to an outdated value, which would overwrite
the clear operation's write to head, thus undoing the clear operation.

Fix race condition in allowAdds that did not account for the non-atomic
"cas" made in appendList, similar to the situation above (tail is set to
outdated value... etc)

Fix many functions not setting the current linked node while iterating,
causing infinite loops.