Guard certain collections against asynchronous access
The guessing game currently for figuring out which plugins cause the
server to crash is to remove plugins and wait for an issue to occur,
which is a long process given the low chance of concurrent modifications
actually killing a server.

This change will at least make server owners know which plugins are
behaving badly, and thus allow them to resolve the issue quicker.