Optimize AreaMap a little bit, implement outsiderange patch better
The add/remove for rectangle methods in AreaMap were not being
inlined due to their size, which is mostly due to the number
of parameters in them, so they were inlined.

AreaMap also now has a size() method, and a getLastViewDistance()
method.

Previously for the isoutsiderange map, we would not allow plugins
to set the radius. Now we do, however we limit it to the max
server view distance (32). We also move the optimization
outside the method, since the JIT would not inline it due
to the fact it could not eliminate the config = false
branch.

The range patch also throws out Mojang's method for tracking
"spawn" chunks and replaces it with a distance map.

This patch also uses an ArrayDeque in ChunkMapDistance since
LinkedList sucks