Retain behaviour of getEntities calls in CraftWorld
In changing to the loaded list, we dropped the isValid check entirely.
Re-introduce the checks isValid used except the chunk check.

Current behaviour should not return dead entities or invalid entities,
and this change ensures that.

Added a note to bring up whether we should return entities with
health < 0 to spigot.