Migrate world config format to paper/spigot format
Now world configuration is under the section `world-settings`, and
defaults are placed in world-settings.default

Worlds will generate their own section, but it will be empty so that
defaults override them if not set.

The previous configuration will migrate to the new format automatically.
World specific settings not overriding default world settings will not be
copied to its new world section, but world specific settings overriding default
world values will be copied. Effectively, this change is not going to
modify the configuration you have set per world. Please verify the config
to verify this before pushing to prod.