Re-add optimise collision checking in player move packet handling
This time, properly check if the move call detected a collision.
Previously, the d11 variable (total distance away from move() call position)
was a bad choice because the y diff variable (d8) used to calculate d11
was clamped to 0 if it was within [-0.5, 0.5] - the right range to hide
falling through a block.