Retain vanilla behavior for error correction rules in collisions
The rules for VoxelShape collision:
- Two shapes only intersect if they are actually more than EPSILON
units into each other.  This also applies to movement checks.

- If the two shapes strictly collide, then the return value
of a collide call will return a value in the opposite
direction of the source move. However, this value
will not be greater in magnitude than EPSILON. Collision
code will automatically round it to 0.

Before this commit, no epsilon values were used - so
the code was subject to floating point errors.

Fixes https://github.com/Spottedleaf/Tuinity/issues/123