Do not return complex parts for entity by class lookup
The class might not match. While complex parts are never added
to chunks (which is why they are special cased), the vanilla
get entities by class does no checking itself, which means
we do not need to.

Add an async catcher to our new getEntitiesByClass method
as well, and change the existing catchers to async catchers.