Rewrite getClosestEntity
No longer make the behavior change. It did not appear to have
a performance benefit. The most likely reason for this is that
typical heavy uses of this function is when there are searches
for entities that fail - so, more search calls are used (once
a target is found, the searching generally stops). Thus the more
heavy uses are repeated failed searches. Instead we use
getSpiralOutChunks so that we call the pathfinder condition less.

Also fix an error in implementation where previous entities from
previous chunk searches would not be wiped from the temp entity list.