Aggressive state checking in region manager
Try to catch state issues right when they occur rather than later

Allow iterators to see new elements added to the iterator safe
ordered reference set while iterating.

Hopefully some of the issues I saw were due to an async usage of
the section iterator, which meant that not all sections in the
region were visible to some of the iterator usages.