Improve behavior for hard colliding entities
Fixes https://github.com/Spottedleaf/Tuinity/issues/249

While this commit closes the above issue, it does
not mean the original problem is completely solved,
it has simply been solved to the best of the
technical abilitites of this game. I expect
this commit however to significantly improve
the collision behavior to the point that we
are much better than vanilla/paper.

There are a couple of existing problems that caused
the server to reject client movements if a player was
colliding with a hard colliding entity:

1. Position desync bewteen client & server

   This was mostly fixed by sending full precision packets
   to clients when updating positions. Due to obvious reasons,
   not even this will fix all of the desync but it will fix
   a lot of it.

2. Hard collision retrieval method returning shapes that were
not colliding

   The hard collision retrieval method was incorrectly returning
   bounding boxes that werent colliding. This caused the
   move packet handling to think that the player was trying
   to move into something it shouldn't be, but in reality
   the player was never doing this. The recent change to patch
   `0035-Optimise-collision-checking-in-player-move-packet-ha.patch`
   removed the collision epsilon there (since it fixed an hclip exploit),
   which then revealed the underlying problem in the retrieval method
   implementation. This method has since been modified so it will
   not return collisions unless they actually collide.