Starlight Handle concurrent chunk generation and lighting better
Thanks to the top tier chunk system, chunk generation
and lighting can occur at the same time. To fix this,
I've adjusted the checkBlock function to no longer
reject updates in chunks that are marked as lit = false,
but still require ChunkStatus.LIGHT. I then adjusted
the block source reading in BlockStarLightEngine to
collect the block sources at once, and if some exception
occurs, try again. It'll eventually get the sources,
and at that point it can light the chunk. If a block source
is added during the lighting, the checkBlock function will
pick it up - so we technically hide the concurrency
problem but make sure lighting is correct in the end,
hopefully.