Fix incorrect status dataconverter for pre 1.13 chunks
Vanilla was setting non-populated OR non-lit chunks to empty, but
really this is just completely wrong. It should be set to "carved"
at minmum, because pre 1.13 chunks went through 3 distinct stages
of generation: carving, population, and lighting - in this order.
There is no "empty" status, because a chunk was simply carved
or it didn't exist. So mapping any chunk data to empty is simply
invalid.

If the chunk is terrain populated, then obviously it must be at
minmum "decorated." If the chunk is lit and populated, then it is marked
"mobs_spawned" (which is what Vanilla is doing, and this is the last
stage before moving to full so it looks correct).

So now here is a table representing the new status conversion:

Chunk is lit        Chunk is populated       Vanilla
     F                       F                empty
     T                       F                empty
     F                       T                empty
     T                       T             mobs_spawned

Chunk is lit        Chunk is populated       Tuinity
     F                       F                carved
     T                       F                carved
     F                       T              decorated
     T                       T             mobs_spawned

This should fix some problems converting old data, as the
changes here are going to prevent the chunk from being regenerated
incorrectly.